requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
matplotlib>=3.8.0
networkx>=3.2.0
reportlab>=4.0.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
numpy>=1.24.0
